Default to the ahash hasher for both std and no_std
The hasher will need to have a default value if we add another
type parameter with a default that goes before the hasher.

IndexMap has the std hasher as default for std and no default
for no_std. It uses hashbrown internally though, which uses
ahash as a default in both cases, do using ahash as a default
is reasonable.

This commit also entirely removes std as a dependency, making
lookupvec actually no_std compatible.